issue #67: adding better perf tests plan
