+++ tests/frame-htmls/perf-async.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+	<title>Performance isolation frame - async</title>
+</head>
+
+<body>
+	<script type="module" src="perf-async.js"></script>
+</body>
+
+</html>
