+++ tests/frame-htmls/perf-sync.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+	<title>Performance isolation frame - sync suite</title>
+</head>
+
+<body>
+	<script type="module" src="perf-sync.js"></script>
+</body>
+
+</html>
